+++ src/Argv/Flags.ts
@@ -1,0 +1,3 @@
+export type Flags <T> = {
+  [P in keyof T]?: string[]
+}