+++ src/Handler/Handler.ts
@@ -1,0 +1,5 @@
+import {Kernel} from '../Kernel'
+
+export interface Handler {
+  execute (kernel: Kernel): Promise<void>
+}