+++ src/Options/OptionsProvider.ts
@@ -1,0 +1,6 @@
+import {Kernel} from '../Kernel'
+import {Partial} from '../Util/Partial'
+
+export interface OptionsProvider <T> {
+  provide (kernel: Kernel): Promise<Partial<T>>
+}