+++ src/Util/Partial.ts
@@ -1,0 +1,3 @@
+export type Partial <T extends Object> = {
+  [P in keyof T]?: T[P]
+}